Extract build_report to a crate
so that it can be called from other skc_* crates
