+++ lib/skc_error/README.md
@@ -1,0 +1,3 @@
+# skc_error
+
+Provides `skc_error::build_report`, a helper to build an error report with the [ariadne](https://crates.io/crates/ariadne) crate.
